multiclass_classification/requirements.txt
openai>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
nltk>=3.6.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.60.0
python-dotenv>=0.19.0